+++ src/main/java/site/offload/offloadserver/db/repository/CharacterMotionRepository.java
@@ -3,5 +3,5 @@ package site.offload.offloadserver.db.repository;
 import org.springframework.data.repository.CrudRepository;
 import site.offload.offloadserver.db.entity.CharacterMotion;
 
-public interface CharacterMotionRepository extends CrudRepository<CharacterMotion, Long> {
+public interface CharacterMotionRepository extends CrudRepository<CharacterMotion, Integer> {
 }
