+++ src/main/java/site/offload/offloadserver/db/repository/VisitedPlaceRepositoiry.java
@@ -3,5 +3,5 @@ package site.offload.offloadserver.db.repository;
 import org.springframework.data.repository.CrudRepository;
 import site.offload.offloadserver.db.entity.VisitedPlace;
 
-public interface VisitedPlaceRepositoiry extends CrudRepository<VisitedPlace, Integer> {
+public interface VisitedPlaceRepositoiry extends CrudRepository<VisitedPlace, Long> {
 }
